requirements.txt
robotframework
Appium-Python-Client
Appium-Flutter-Finder
requests
decorator
selenium
kitchen
six